requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
mysql-connector-python==9.0.0
python-dotenv==1.0.1
Werkzeug==3.0.3
Flask-SQLAlchemy==3.1.1
Pillow==10.4.0

